Move query base class into seperate file
