+++ benchmark_suite/include/moveit_benchmark_suite/dataset.h
@@ -54,10 +54,10 @@
 
 #include <moveit_benchmark_suite/io.h>
 
+#include <moveit_benchmark_suite/query.h>
+
 namespace moveit_benchmark_suite
 {
-MOVEIT_CLASS_FORWARD(Query);
-MOVEIT_CLASS_FORWARD(Result);
 MOVEIT_CLASS_FORWARD(Data);
 MOVEIT_CLASS_FORWARD(DataSet);
 
@@ -89,57 +89,6 @@ using Metric = boost::variant<bool, double, int, std::size_t, std::string>;
  */
 std::string toMetricString(const Metric& metric);
 double toMetricDouble(const Metric& metric);
-
-using QueryGroup = std::string;
-using QueryName = std::string;
-using QueryResource = std::string;
-
-using QueryGroupName = std::map<QueryGroup, QueryName>;
-
-struct Query
-{
-  /** \brief Empty constructor.
-   */
-  Query() = default;
-
-  virtual ~Query(){};
-
-  Query(const std::string& name, const QueryGroupName& group_name_map) : name(name), group_name_map(group_name_map){};
-
-  std::string name;  ///< Name of this query.
-  QueryGroupName group_name_map;
-};
-
-// pair - wise combinations of a query
-struct QuerySetup
-{
-  /** \brief Empty constructor.
-   */
-  QuerySetup() = default;
-
-  void addQuery(const QueryGroup& group, const QueryName& name, const QueryResource& resource)
-  {
-    auto it = query_setup.find(group);
-    if (it == query_setup.end())
-      query_setup.insert(std::pair<QueryGroup, std::map<QueryName, QueryResource>>(group, { { name, resource } }));
-    else
-      it->second.insert(std::pair<QueryName, QueryResource>(name, resource));
-  }
-
-  std::map<QueryGroup, std::map<QueryName, QueryResource>> query_setup;
-};
-
-class Result
-{
-public:
-  Result() = default;
-
-  virtual ~Result(){};
-  /** \name Planning Query and Response
-      \{ */
-
-  bool success = false;  ///< Was the plan successful?
-};
 
 /** */
 class Data
